Move LockfileContent to its own module
